--- components/index.js
@@ -1,5 +1,0 @@
-import iobioCharts from 'iobio-charts';
-
-const { createHistogram, createPercentBox, HistogramElement, PercentBoxElement } = iobioCharts;
-
-export { createHistogram, createPercentBox, HistogramElement, PercentBoxElement };
